backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for chat app"
requires-python = ">=3.10"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "fastapi==0.121.3",
    "h11==0.16.0",
    "idna==3.11",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "sniffio==1.3.1",
    "starlette==0.50.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = []






